renamed IOptionHandler to CliOptionHandler, renamed AlphaConfig to SystemConfig, renamed AlphaContext to AlphaConfig
