+++ src/main/java/at/ac/tuwien/kr/alpha/Main.java
@@ -40,7 +40,7 @@ import org.slf4j.LoggerFactory;
 
 import at.ac.tuwien.kr.alpha.common.AnswerSet;
 import at.ac.tuwien.kr.alpha.common.Program;
-import at.ac.tuwien.kr.alpha.config.AlphaContext;
+import at.ac.tuwien.kr.alpha.config.AlphaConfig;
 import at.ac.tuwien.kr.alpha.config.CommandLineParser;
 import at.ac.tuwien.kr.alpha.config.InputConfig;
 import at.ac.tuwien.kr.alpha.solver.Solver;
@@ -57,7 +57,7 @@ public class Main {
 
 	public static void main(String[] args) {
 		CommandLineParser commandLineParser = new CommandLineParser(Main.ALPHA_CALL_SYNTAX, (msg) -> Main.exitWithMessage(msg, 0));
-		AlphaContext ctx = null;
+		AlphaConfig ctx = null;
 		try {
 			ctx = commandLineParser.parseCommandLine(args);
 		} catch (ParseException ex) {
