+++ src/main/java/at/ac/tuwien/kr/alpha/config/CliOptionHandler.java
@@ -4,7 +4,7 @@ import org.apache.commons.cli.Option;
 import org.apache.commons.cli.ParseException;
 
 @FunctionalInterface
-public interface IOptionHandler<T> {
+public interface CliOptionHandler<T> {
 
 	void handleOption(Option opt, T dest) throws ParseException;
